Make getReadableDate a static util method
